+++ src/user/rd.cpp
@@ -134,5 +134,5 @@ size_t __stdcall rd(const kiv_os::TRegisters &regs) {
 	std::vector<std::string> args = kiv_os_lib::getArgs("rd", regs);
 	std::vector<char *> argv = kiv_os_lib::getArgsDataPointer(args);
 
-	return rd_program::rd_main(argv.size(), argv.data());
+	return rd_program::rd_main(static_cast<int>(argv.size()), argv.data());
 }
