+++ src/user/sort.cpp
@@ -38,5 +38,5 @@ size_t __stdcall sort(const kiv_os::TRegisters &regs) {
 	std::vector<std::string> args = kiv_os_lib::getArgs("sort", regs);
 	std::vector<char *> argv = kiv_os_lib::getArgsDataPointer(args);
 
-	return sort_program::sort_main(argv.size(), argv.data());
+	return sort_program::sort_main(static_cast<int>(argv.size()), argv.data());
 }
